fix: move var functions to separate file & namespace
BREAKING CHANGE: this changes the namespace of `is_closed_resource()` and `is_zero()`
