+++ src/constants.php
@@ -15,8 +15,3 @@ namespace empaphy\usephul;
  * This mask matches every possible past and future PHP error level.
  */
 const E_EVERYTHING = 0x7FFFFFFF;
-
-/**
- * Default error tolerance.
- */
-const PHP_ZERO_TOLERANCE = 0.00000000001;
